Code
Tic-tac-toe is a pencil-and-paper game for two players, X and O, who
take turns marking the spaces in a 3×3 grid. The player who succeeds in
placing three respective marks in a horizontal, vertical, or diagonal
row wins the game.

Now you are going to write code to play Tic-tac-toe with others or
computer bot.

Input
The input will be an 3x3 matrix consisting only of 0, 1 or 2. Then
another line will follow which will contain a number 1 or 2, which is
your player id.

In the given matrix, top-left is [0,0] and bottom-right is [2,2]. The
x-coordinate increases from left to right, and y-coordinate increases
from top to bottom.

The cell marked 0 means it doesn't contain any marks. The cell marked 1
means it contains player 1's mark which is gray in color. The cell
marked 2 means it contains player 2's mark which is black in color.

Output
Print the coordinates of the cell (x, y) where you want to play your
move.

Starting state
0 0 0
0 0 0
0 0 0